Feature/reprocessing (#25)
* change dtype from bool to uint8

* specify saving dtype

* pass percentile for height_percentile

* load image once, threshold multiple times

* clean up renaming

* update configs

* add simple consistency validation

* update steps

* update configs

* bump cytodl version

---------

+++ morflowgenesis/steps/alignment.py
@@ -56,7 +56,7 @@ def align(
         output_type="image",
         image_id=image_object.id,
     )
-    output.save(new_seg)
+    output.save(new_seg.astype(np.uint16))
     image_object.add_step_output(output)
     image_object.save()
